feat: change grade type from Integer to Float across models and update related logic

+++ server/src/main/java/br/ufal/ic/odontolog/dtos/ReviewableSubmitSupervisorReviewDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class ReviewableSubmitSupervisorReviewDTO {
   @NotNull private String comments;
 
-  @NotNull private Integer grade;
+  @NotNull private Float grade;
 
   @NotNull private Boolean approved;
 }
