+++ server/src/main/java/br/ufal/ic/odontolog/models/TreatmentPlan.java
@@ -85,7 +85,7 @@ public class TreatmentPlan extends Reviewable {
 
   @Override
   public void submitSupervisorReview(
-      Supervisor supervisor, String comments, Integer grade, Boolean approved) {
+      Supervisor supervisor, String comments, Float grade, Boolean approved) {
     this.getState().submitSupervisorReview(this, supervisor, comments, grade, approved);
   }
